// projects/ui/src/lib/components/po-chart/interfaces/po-chart-min-max-values.interface.ts
/**
 * @docsPrivate
 *
 * @usedBy PoChartComponent
 *
 * @description
 *
 * *Interface* `PoChartMinMax` para especificação de valores específicos das séries.
 */
export interface PoChartMinMaxValues {
  /**
   * O valor mínimo das séries.
   */
  minValue?: number;

  /**
   * O valor máximo das séries
   */
  maxValue?: number;
}